requirements.txt
python==3.9.7
GitPython==3.1.27
